requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
googlemaps==4.10.0
schedule==1.2.0
python-dotenv==1.0.0
pdfplumber==0.11.0
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0
pillow==10.4.0

